requirements.txt
django
progressbar2
libact
feedparser
newspaper3k
pusher
sklearn
scipy
git+https://github.com/zWaR/PythonPusherClient.git
django-dotenv
django-after-response
after_response
requests
